video-conferencing-app/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
httpx==0.24.1
python-dotenv==1.0.0
firebase-admin==6.1.0  # Only if using Firebase Admin SDK